+++ src/Int/IntIdInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CyrilVerloop\DoctrineProperties\Int;
+
+/**
+ * An interface for integer id.
+ *
+ * @psalm-api
+ */
+interface IntIdInterface
+{
+    // Properties :
+
+    public ?int $id { get; }
+}