+++ CHANGELOG.md
@@ -6,3 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- boolean properties : Active, Available.
+- integer properties : IntId, Priority.
+- string properties: Description, Email, Firstname, Lastname, Name, Slug, Surname.
+- an abstract class with IntId primary key : AbstractIntId.